+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Live DevOps Pipelines enhancements
   - Display Git & ticketing providers icons
   - Add button to view pipeline configuration from pipeline view
+  - Allow more formats of jiraHost
 
 ## [6.14.5] 2024-10-22